bio_agent/requirements.txt
psycopg2-binary>=2.9.9
boto3>=1.34.0
h3>=3.7.0
